+++ Admin/RouteAdmin.php
@@ -66,16 +66,18 @@ class RouteAdmin extends Admin
                 array('choice_list' => array(), 'select_root_node' => true, 'root_node' => $this->routeRoot)
             )
             ->add('name', 'text')
-            ->add('addFormatPattern', null, array('required' => false, 'help' => 'form.help_add_format_pattern'))
-            ->add('addTrailingSlash', null, array('required' => false, 'help' => 'form.help_add_trailing_slash'))
         ->end();
 
         if (null === $this->getParentFieldDescription()) {
             $formMapper
                 ->with('form.group_general')
-                ->add('variablePattern', 'text', array('required' => false))
-                ->add('content', 'doctrine_phpcr_odm_tree', array('choice_list' => array(), 'required' => false, 'root_node' => $this->contentRoot))
-                ->add('defaults', 'sonata_type_immutable_array', array('keys' => $this->configureFieldsForDefaults()))
+                    ->add('content', 'doctrine_phpcr_odm_tree', array('choice_list' => array(), 'required' => false, 'root_node' => $this->contentRoot))
+                ->end()
+                ->with('form.group_advanced')
+                    ->add('variablePattern', 'text', array('required' => false), array('help' => 'form.help_variable_pattern'))
+                    ->add('defaults', 'sonata_type_immutable_array', array('keys' => $this->configureFieldsForDefaults()))
+                    ->add('options', 'sonata_type_immutable_array', array('keys' => $this->configureFieldsForOptions()), array('help' => 'form.help_options'))
+                ->end()
             ->end();
         }
     }
@@ -83,7 +85,7 @@ class RouteAdmin extends Admin
     protected function configureDatagridFilters(DatagridMapper $datagridMapper)
     {
         $datagridMapper
-            ->add('name', 'doctrine_phpcr_string');
+            ->add('name', 'doctrine_phpcr_nodename');
     }
 
     public function setRouteRoot($routeRoot)
@@ -110,7 +112,7 @@ class RouteAdmin extends Admin
 
     protected function configureFieldsForDefaults()
     {
-        $defaults =  array(
+        $defaults = array(
             '_controller' => array('_controller', 'text', array('required' => false)),
             '_template' => array('_template', 'text', array('required' => false)),
             'type' => array('type', 'cmf_routing_route_type', array(
@@ -125,7 +127,7 @@ class RouteAdmin extends Admin
                 $defaults[$name] = array($name, 'text', array('required' => false));
             }
         }
-        
+
         //parse variable pattern and add defaults for it - taken from routecompiler
         /** @var $route Route */
         $route =  $this->subject;
@@ -140,6 +142,24 @@ class RouteAdmin extends Admin
         }
 
         return $defaults;
+    }
+
+    protected function configureFieldsForOptions()
+    {
+        $options = array(
+            array('add_locale_pattern', 'checkbox', array('required' => false, 'label' => 'form.label_add_locale_pattern', 'translation_domain' => $this->translationDomain)),
+            array('add_format_pattern', 'checkbox', array('required' => false, 'label' => 'form.label_add_format_pattern', 'translation_domain' => $this->translationDomain)),
+            array('add_trailing_slash', 'checkbox', array('required' => false, 'label' => 'form.label_add_trailing_slash', 'translation_domain' => $this->translationDomain)),
+        );
+
+        $dynamicOptions = $this->getSubject()->getOptions();
+        foreach ($dynamicOptions as $name => $value) {
+            if (!isset($options[$name])) {
+                $options[$name] = array($name, 'text', array('required' => false));
+            }
+        }
+
+        return $options;
     }
 
     public function prePersist($object)