+++ Doctrine/Phpcr/IdPrefixListener.php
@@ -15,30 +15,41 @@ namespace Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr;
 use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
 
 /**
- * Doctrine PHPCR-ODM listener to set the idPrefix on routes
+ * Doctrine PHPCR-ODM listener to tell routes what part of their id is the URL.
+ *
+ * This listener knows about the RouteProvider and uses its prefixes to
+ * identify routes that could need the prefix. In case prefixes overlap, the
+ * order matters as the first matching prefix is taken.
  *
  * @author David Buchmann <mail@davidbu.ch>
  */
 class IdPrefixListener
 {
     /**
-     * The prefix to add to the url to create the repository path
+     * Used to ask for the possible prefixes to remove from the repository ID
+     * to create the URL.
      *
-     * @var string
+     * @var PrefixCandidates
      */
-    protected $idPrefix = '';
+    protected $candidates;
 
-    public function __construct($prefix)
+    /**
+     * This listener only makes sense together with the PrefixCandidates
+     * strategy.
+     *
+     * @param PrefixCandidates $candidates
+     */
+    public function __construct(PrefixCandidates $candidates)
     {
-        $this->idPrefix = $prefix;
+        $this->candidates = $candidates;
     }
 
     /**
-     * @param $prefix
+     * @return array
      */
-    public function setPrefix($prefix)
+    protected function getPrefixes()
     {
-        $this->idPrefix = $prefix;
+        return $this->candidates->getPrefixes();
     }
 
     public function postLoad(LifecycleEventArgs $args)
@@ -51,16 +62,24 @@ class IdPrefixListener
         $this->updateId($args);
     }
 
+    public function postMove(LifecycleEventArgs $args)
+    {
+        $this->updateId($args);
+    }
+
     protected function updateId(LifecycleEventArgs $args)
     {
         $doc = $args->getObject();
 
         // only update route objects and only if the prefix can match, to allow
         // for more than one listener and more than one route root
-        if (($doc instanceof PrefixInterface)
-            && ! strncmp($this->idPrefix, $doc->getId(), strlen($this->idPrefix))
-        ) {
-            $doc->setPrefix($this->idPrefix);
+        if ($doc instanceof PrefixInterface) {
+            foreach ($this->getPrefixes() as $prefix) {
+                if (! strncmp($prefix, $doc->getId(), strlen($prefix))) {
+                    $doc->setPrefix($prefix);
+                    break;
+                }
+            }
         }
     }
 }